requirements.txt
# 博客生成脚本依赖

# GitHub API 客户端
PyGithub>=2.1.1

# RSS/Atom feed 生成
feedgen>=0.9.0

# Markdown 解析（GitHub Flavored Markdown）
marko>=1.2.0

# XML 处理（用于 RSS 生成）
lxml>=4.9.0

# requests 库（PyGithub 依赖）
requests>=2.31.0
